Added basket

+++ webshop/browse.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, request, g
 import MySQLdb as mdb
-
+from flask.ext.login import current_user
 from db_utils import get_all_categories
 
 browse_page = Blueprint('browse_page', __name__, template_folder='templates')
@@ -52,23 +52,24 @@ def display_category(category_id=None, order_by='name', order='asc', error_messa
 @browse_page.route('/browse/update_asset/<category_id>', defaults={'asset_id': None}, methods=['POST'])
 @browse_page.route('/browse/update_asset/<category_id>/<asset_id>', methods=['POST'])
 def update_asset(category_id, asset_id):
-    db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
-    db.execute('update Asset set '
-               'name=%s, '
-               'price=%s, '
-               'amount=%s, '
-               'imagePath=%s, '
-               'Category_idCategory=%s '
-               'where idAsset=%s',
-               [
-                   request.form['text-name'],
-                   request.form['text-price'],
-                   request.form['text-amount'],
-                   request.form['text-imagePath'],
-                   request.form['select-category'],
-                   asset_id
-               ])
-    db.connection.commit()
+    if current_user.admin:
+        db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
+        db.execute('update Asset set '
+                   'name=%s, '
+                   'price=%s, '
+                   'amount=%s, '
+                   'imagePath=%s, '
+                   'Category_idCategory=%s '
+                   'where idAsset=%s',
+                   [
+                       request.form['text-name'],
+                       request.form['text-price'],
+                       request.form['text-amount'],
+                       request.form['text-imagePath'],
+                       request.form['select-category'],
+                       asset_id
+                   ])
+        db.connection.commit()
 
     return display_category(category_id)
 
@@ -76,52 +77,52 @@ def update_asset(category_id, asset_id):
 @browse_page.route('/browse/delete_asset/<category_id>', defaults={'asset_id': None}, methods=['POST'])
 @browse_page.route('/browse/delete_asset/<category_id>/<asset_id>', methods=['POST'])
 def delete_asset(category_id, asset_id=None):
-    # TODO: admin
-    db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
-    db.execute('delete from Asset where idAsset=(%s)', [asset_id])
-    db.connection.commit()
+    if current_user.admin:
+        db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
+        db.execute('delete from Asset where idAsset=(%s)', [asset_id])
+        db.connection.commit()
     return display_category(category_id)
 
 
 @browse_page.route('/browse/delete_category/<category_id>', methods=['POST'])
 def delete_category(category_id):
-    # TODO: admin
-    try:
-        db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
-        db.execute('delete from Category where idCategory=(%s)', [category_id])
-        db.connection.commit()
-    except mdb.IntegrityError as e:
-        return display_category(category_id, error_message="Category must be empty before removal.")
+    if current_user.admin:
+        try:
+            db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
+            db.execute('delete from Category where idCategory=(%s)', [category_id])
+            db.connection.commit()
+        except mdb.IntegrityError as e:
+            return display_category(category_id, error_message="Category must be empty before removal.")
     return display_category()
 
 
 @browse_page.route('/browse/add_asset/<category_id>', methods=['POST'])
 def add_asset(category_id):
-    # TODO: admin
-    db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
-    db.execute('insert into Asset '
-               '(name, price, amount, imagePath, Category_idCategory)'
-               ' values '
-               '(%s, %s, %s, %s, %s)',
-               [
-                   request.form['text-name'],
-                   request.form['text-price'],
-                   request.form['text-amount'],
-                   request.form['text-imagePath'],
-                   request.form['select-category']
-               ])
-    db.connection.commit()
+    if current_user.admin:
+        db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
+        db.execute('insert into Asset '
+                   '(name, price, amount, imagePath, Category_idCategory)'
+                   ' values '
+                   '(%s, %s, %s, %s, %s)',
+                   [
+                       request.form['text-name'],
+                       request.form['text-price'],
+                       request.form['text-amount'],
+                       request.form['text-imagePath'],
+                       request.form['select-category']
+                   ])
+        db.connection.commit()
     return display_category(category_id)
 
 
 @browse_page.route('/browse/add_category/<category_id>', methods=['POST'])
 def add_category(category_id):
-    # TODO: admin
-    if len(request.form['text-category_name']) > 0:
-        db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
-        db.execute('insert into Category (name) values (%s)', [request.form['text-category_name']])
-        db.connection.commit()
-        db.execute('select * from Category order by idCategory desc limit 1')
-        return display_category(str(db.fetchall()[0]['idCategory']))
+    if current_user.admin:
+        if len(request.form['text-category_name']) > 0:
+            db = getattr(g, 'db', None).cursor(mdb.cursors.DictCursor)
+            db.execute('insert into Category (name) values (%s)', [request.form['text-category_name']])
+            db.connection.commit()
+            db.execute('select * from Category order by idCategory desc limit 1')
+            return display_category(str(db.fetchall()[0]['idCategory']))
 
     return display_category(category_id)
